Modificação na estrutura do programa.
